+++ packages/react/src/atomic/Button/src/index.ts
@@ -1,5 +1,5 @@
 import { Button } from "./Button";
 
 export { Button } from "./Button";
-export type { ButtonProps } from "./button.types";
+export type { ButtonProps } from "./Button";
 export default Button;