+++ packages/react/src/atomic/Button/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The Button component allows us to initiate actions, make state or page changes.
 
+## 2023-03-24 `2.5.0`
+
+### 💡 Others
+
+- Adjusted the component properties interface to use `ComponentPropsWithRef` and include the `as` prop in the output interface. ([#132](https://github.com/TiendaNube/nimbus-design-system/pull/132) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2023-02-23 `2.4.0`
 
 ### 🎉 New features